backend/requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Cors==4.0.0
flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
python-engineio==4.8.1
python-socketio==5.10.0